Regularizing selector (and request) interface. Modeling note included.
Recall that selector is a special type of request. (Selector inherits from Request.)

The following are abstract methods:

    Request._get_payload(score_specification, voice_name)

    Selector._get_payload(score_specification, voice_name)
    Selector._get_timespan(score_specification, voice_name)
    Selector._get_timespan_and_payload(score_specification, voice_name)

MODEL: Request models a request *for something*.
It is the responsibility of Request._get_payload() to return that something.
Important to understand here is that the something returned by Request._get_payload()
can be thought of as existing *outside of time* (in Xenakis's sense).
For example, AbsoluteRequest._get_payload() probably returns a list of numbers, pairs or other constants.
This list will eventually be placed in the score. But the list itself has no (score-relative) timespan.
Because Request._get_payload() can not be guaranteed to define a score-relative timespan, 
Request implements _get_payload() but does not impelment either _get_timespan() or _get_timespan_and_payload().

MODEL: Selector models a request for something *in the score*.
BeatSelector._get_payload() returns beats *in the score*. This means that the beats
are guaranteed to subtend a score-relative timespan (once interpretation reaches a certain point).
It is the job of Selector._get_timespan() to return this score-relative timespan.

MODEL: Selector._get_payload() and Selector._get_timespan() wrap Selector._get_timespan_and_payload().