Considerations on preparing for payload-carrying timespan class. Modeling update included.
This commit regularizes all RegionProduct initializers.

The region product initializers are these:

    RegionProduct.__init__(payload=None, voice_name=None, timespan=None)
    DivisionRegionProduct.__init__(payload=None, voice_name=None, timespan=None)
    RhythmRegionProduct.__init__(payload=None, voice_name=None, timespan=None)

So perhaps RegionProduct will subclass the payload-carrying timespan class and extend with 'voice_name'.

MODEL: it's possible that we have one last (mercifully easier) case of aggregation / inheritance
to think through when it comes to the idea of the payload-carrying timespan class. We could indeed
implement a PayloadCarryingTimespan class. (And I think this is way we've been leaning recently.)
But pause to consider that a discovery we've made is that most of the classes we've had inherit
from (pure) Timespan have had to be refactored to *have* a timespan instead. I think it might
be the same with the idea of payload-carrying timespan classes. Why? Because there are going
to be many classes (like BreakpointFunction, as pointed out in the recent commits .... and like
the RegionProduct classes updated in this commit) that will have both a payload and a timespan.
But maybe it's smarter to simply speak of those classes "implementing the payload / timespan interface",
to use the usual Python way of talking about things. (Think about the Python concept of an itertor:
classses that 'are' iterators don't have to inherit from anything; they just need to implement
next(), __iter__() and raise StopIteration at the right time; "if it walks like a duck ...".)

So on second thought we'll hold off on implementing any type of payload-carrying timespan *class*
for the moment.