Renamed CommandRequest classes to SettingLookupRequest classes.
This comes very close (finally) to disambiguating the use of 'command' in the system.

Where this is going is that 'command' will be used to refer only to
an interpreter byproduct never seen by the composer.

(Recall that during interpretation: setting > command > product. 
Command and product are never seen by the composer.)