Extended ScoreSpecification.get_region_commands_for_voice().
Recall that the method changes settings into commands.

Previously the method did not check to make sure that each
setting-in-segment correponded to a timespan that actually exists for that segment.
(This occurs when a setting exists that says to do something for measure
10 in a segment that has only 9 measures.)
The method now checks to make sure that each setting-in-segment
corresponds to a timespan that actually exists in that segment.