Added new developer script tool:
    * developerscripttools.TestAndRebuildScript

Run the script on the commandline via `ajv rebuild`.

The script runs `py.test -x` on the current directory.  If all tests pass, it then runs the doctests on all modules in the current directory.  If all doctests pass, it finally rebuild the experimental and mainline docs.  If any tests fail, it will not proceed to the next step.

It will bail on the first failing test and first failing doctest.