Updated RhythmRegionProduct docstrings.
This is part of the work necessary to implement counttime component selector payload callbacks.

The reason is that the counttime components selected by a counttime component selector
are the components of the payload of one or more rhythm region products.

That is, it is rhythm region product payloads that are selected by
counttime component selectors.