Added Infinity and NegativeInfinity to built-in namespace.
Just two one-line definitions in abjad/tools/__init__.py file to alias to Python's float-based infinity constructs:

    __builtins__['Infinity'] = float('infinity')
    __builtins__['NegativeInfinity'] = float('-infinity')

Examples:

    >>> NegativeInfinity < Infinity
    True

    >>> NegativeInfinity < -1000000
    True

    >>> 99999999 < Infinity
    True

    >>> 2**38 < Infinity
    True


Infinity and NegativeInfinity are not datatstructuretools.OrdinalConstant objects.

The reason for this is that ordinal constants are specifically designed *not* to compare to numbers
(but only to each other). 

This means that expressions like 'Ininify < Right' do not evaluate.