TimespanExpression no longer inherits from Timespan. Modeling update included.
MODEL: The model is now rapidly converging to the point where the system
implements a number of different expression classes, all inheriting from Expression.
The endpoint of this work is that all expression classes will implement
a single _evaluate(score_specification) method. This means
that we will finally be able to speak of "evaluating expressions" in the usual way.
It will also be the case that all of our expressions will just happen to be
implemented as full-fledged object-oriented classes.

With the above in mind it is easy to see that TimespanExpression shouldn't
inherit from Timespan. TimespanExpression should instead *evaluate to* a timespan.

Note that this bit of convergence in the model will be complete when
all _get_payload(), _get_timespan() and _get_payload_and_timespan() methods
have been replaced with an _evaluate() method.