Selector no longer inherits from TimespanExpression. Modeling update included.
MODEL: This is a major change and a major clarification. The selector
classes now inherit from Expression. Why? Because the idea of
something like a BeatSelector object is basically a frozen request
for a bunch of (contiguous) beats. While it is true that the expression
will evaluate to a group of beats that will carry a timespan,
the primary meaning of the BeatSelector expression is to pick out beats;
the timespan attribute can now be referenced with dot-notation like this:

    beats = red_segment.select_beats('Voice 1').timespan
    
A TimespanExpression is returned with the complete TimespanCallbackMixin
interface available.

This will greatly clean up evaluation because selectors will soon
evaluate to a payload *only* and will not evaluate to a timespan.
(Rather the selector.timespan TimespanExpression will evaluate
to a timespan.)

This will also remove the presence of two separate callback stacks
on any one class. That is, every expression class will soon feature
one and only one callback stack.