Renamed SettingLookupRequest. The new name is SettingLookup. Modeling update included.
Same for the other lookup classes.

MODEL: the idea of a "lookup request" is a bit of a misnomer.
Going forward only "request" is to be reserved for only server requests.
These are to be the only request classes in the system:
    HandlerRequest
    StatalServerRequest
These are not really implemented yet. But when implemented these
will be the object-oriented models of calls against a server.