Added stub StartPositionedObject class.
The class will serve to model things (like the RegionProduct classes)
that are positioned at some absolute offset in the score
(and that *are not* anchored to some symbolic timespan in the score).

MODEL: notionally, anchored objects are dealt with in the first half
of interpretation; start-positioned objects are created in the second
half of interpretation.