Payload callbacks only ever apply to payload expressions.
MODEL: This is a major point of clarification in the model.
You can now glance at a list of the class names in the system and
tell exactly which ones may be the target of payload callback
application and which can not.

MODEL: Notice that there is a crucial difference between the
classes that *implement* the PayloadCallbackMixin and the
classes that may be the *target* or payload callback application.
The classes that implement the PayloadCallbackMixin are
the *agents* of payload callback application; it is these
classes that call _apply_callbacks() at some point during
interpretation. There is a hard division between classes
that act as the agents of payload callback application
and classes that act as the targets of payload callback application.
Every class that implements the PayloadCallbackMixin acts
as the agent of payload callback application and no class
that implements the PayloadCallbackMixin acts as the target
of payload callback application.