Parallel SetExpression, AnchoredSetExpression classes now exist.
The purpose in the split is to model the difference between input set expressions
and timespan-scoped set expressions. What is the difference between input
set expressions and timespan-scoped set expressions? Input set expressions
are allowed to have an expression target timespan; but the entire point of
timespan-scoped set expressions is that timespan-scoped set expressions
have a literal target timespan. This means that input set expressions are anchored
while timespan-scoped set expressions are not.