Extend ScoreSpecificationInterface with start-offset lookups.
Example. You can now say this:

    lookup = score_specification.timespan.start_offset.look_up_rhythm_set_expression('Voice 1')
    blue_segment.set_rhythm(lookup)

Interpretation will then use the start-offset of the entire score as the basis for the rhythm lookup.

This is a structural equivalent to asking for the start-offset of the first segment in score
without having to know anything about the first segment in score.

Test and PDF included.