Extend ScoreSpecificationInterface with stop-offset lookups.
Example. You can now say this:

    offset = score_specification.timespan.stop_offset.translate(Duration(-1, 8))
    lookup = offset.look_up_rhythm_set_expression('Voice 1')
    red_segment.set_rhythm(lookup)

Interpretation will then use the stop-offset of the entire score as the basis for the rhythm lookup.

This is a structural equivalent to asking for the stop-offset of the last segment in score
without having to know anything about the last segment in score.

Test and PDF included.