Added first tests to show score-anchored set expressions.
Example. This now works:

    score_specification.set_rhythm(library.sixteenths)

This also now works:

    score_specification.set_divisions([(3, 16)])

Even this now works:

    score_specification.set_time_signatures([(1, 8), (2, 8), (3, 8)])

Note that the functionality is currently implemented at only the level of a foothold.

You still need to create a (dummy) segment specification in order for the examples above to work.
The reason for this is that interpretation is currently treating score-anchored settings
as anchored to the first segment in score. Future commits will remove this constraint.

This commit also includes the first examples of segment selection.

Example. This now works:

    segments = score_specification.select_segments('Voice 1')[2:4]

Even this now works:

    segments.timespan.set_rhythm(library.sixteenths)

PDFs included.