Integrating score-anchored expressions.
Recall that anchored expressions can anchor to one of three things:

    * to a segment
    * to another anchored expression
    * to the score

The work right now is working on the last of these three cases.

As of this commit score-anchored expressions now refuse calls to copy_set_expression_to_segment_name().

This is the desired behavior. Setting-anchored expressions can be reanchored to another segment;
score-anchored expressions should not be renanchored to anything.