Introducing the formal model of expression 'rootedness'. Modeling update included.
As of this commit AnchoredExpression now implements three new read-only properties:

    * AnchoredExpression.is_score_rooted
    * AnchoredExpression.is_segment_rooted
    * AnchoredExpression.root

The twin concept of 'anchor' and 'root' work together.

Recall that anchored expressions can anchor to one of THREE things:

    * to a segment
    * to another anchored expression
    * to the score

Now notice that anchored expressions can be rooted to one of only TWO things:

    * to a segment
    * to the score

MODEL: How do the anchor and root of an anchored expression differ? 
The anchor of an anchored expression is *the next* thing an anchored expression grabs onto. 
The root of an anchored expression is *the last* thing that the entire chain 
of a (possibly arbitrarily composed) expression grabs onto.

MODEL: *This means that the root is the recursive base case in compound expressions.*