Improved lineage diagram output in the docs.
The GraphvizGraph output of InheritanceDiagrams is now pickled and stored with the appropriate doctree nodes during doc building, rather than the GraphvizGraph graphviz_format.  This means that when docs are written to a particular output format (LaTeX vs HTML), the unpickled GraphvizGraph instance can be hand-tailored to better fit that output format.

In practice, the LaTeX diagrams are constrained to a 6.5"x6" box with smaller font sizes.  The HTML diagrams have no page-size constraint, as they will be downsized by ImageMagick's convert in a later step.