Removed Spanner __deepcopy__ definition.
CHANGE: must be noted in the 2.12 change log.

MODEL: the conceptual take-away is that Abjad components no longer override deepcopy.
What does this do? This means, for example, that deepcopying a single note
copies the entire score in which the note is included. (And a comparison to
basic data structures would be to notice that Abjad score trees
are rich in references. This means that deepcopying an element from an Abjad
score tree now has much the same effect as deepcopying an element from
a doubly linked list.)