Implementing pitch interpretation.
Renamed AnchoredSetExpression.
The new name is TimeContiguousAnchoredSetExpression.

The reason for the rename is that implementation is proceeding to
arbtrary selections over a collection of possibly discontiguous timespans.
Everything implemented in spectools so far assumes that select expressions
pick out timespan-contiguous parts of score. But this will soon change.
And the pending change means that AnchoredSetExpression now has to be renamed
to the more specification TimeContiguousAnchoredSetExpression.