Implementing pitch interpretation.
Renamed ArbitraryTargetPitchSetExpression.
The new name is GeneralizedPitchSetExpression.

What's going on here is that the word 'generalized' is being
worked into a number of the expression classes. Generalized
expressions wil be those expressions that can handle
time-contiguous or time-discontiguous expressions.
It's likely that once generalized expressions work
the (unmarked) nongeneralized expressions will simply
be replaced by the generalized expressions. But this will
take some time.