Implementing pitch interpretation.
Extended CounttimeComponentSelectExpression with special __add__() method.

Example:

    first_leaves = score_specification.select_leaves('Voice 1')[:10]
    last_leaves = score_specification.select_leaves('Voice 2')[-10:]
    leaves = first_leaves + last_leaves

The 'leaves' expression that results is a SelectExpressionInventory
(which generalizes a single select expression to a collection of arbitrary select expressions).

Notice that there are very few constraints placed on the number and type of
counttime component select expressions that can be added together:
different voices are allowed, discontiguous timespans are allowed, overlapping
timespans are allowed.

The reason for the flexibility is that what's being model is an arbitrary
selection of notes or chords to be pitched.