Preparing statal servers for use.
Renamed PayloadExpression. The new name is IterablePayloadExpression.
This clears the way for the (re)implementation of an (unqualified) PayloadExpression.
This is important to do because the current implementation of PayloadExpression
is ambiguous as to whether the class models a wrapper around a constant piece of 
arbitrary information or a wrapper around a constant piece of iterable information.