Cleaning up specificationtools package architecture.
The goal of the next several commits is to move everything into the specificationtools package.
This will remove the expressiontools package altogether.

The commits will stage the process by first doing the reverse and migrating the
(smaller number of) specificationtools classes into expressiontools
and then renaming expressiontools as specificationtools.

Classes migrated from specificationtools to expressiontools in this commit:

    ContextDictionary
    ContextProxy
    VoiceDictionary
    VoiceProxy