Extending pitch interpretation.
Implemented composer set_aggregate() method.

Example:

    score_specification.select_leaves('Voice 1').set_pitches(library.example_pitches_1())
    score_specification.select_leaves('Voice 1').set_aggregate(library.example_aggregates[0])

The example sets pitches on all leaves in voice 1 and then registers each 
pitch according to library.example_aggregates[0].

The example aggregates added to the library are these:

    example_aggregates = [
        [-13, -5, 2, 3, 6, 8, 10, 12, 13, 16, 17, 21],
        [-13, -12, -10, -5, -2, 5, 6, 8, 15, 16, 21, 25],
        [-22, -14, -12, -8, -4, -1, 7, 17, 18, 21, 27, 37],
        [-2, 0, 2, 7, 8, 11, 15, 16, 17, 18, 21, 25]]

Notice that each is a twelve-note chord.

This means that pitch-class 0 in the example is registered as 12,
pitch-class 1 as 13, pitch-class 2 as 2, pitch-class 3 as 3, and so on.

PDF included.