Extended iotools.get_last_output_file_name() and iotools.get_next_output_file_name().
The functions now taken an optional path=None keyword.

The functions read the Abjad output directory when path is none.

The functions read any arbitrary directory when path is not none.

This makes it possible to use these functions on composer-specific diretories when working on score.

This change is small enough that it doesn't need to be included in the 2.12 release notes.