Added test to show generalized grob override (and revert) specification.
This comes for free because spanners have a grob override proxy.

Example:

    divisions = score_specification.select_divisions('Voice 1')[2:4]
    leaves = divisions.timespan.select_leaves('Voice 1')
    spanner = spannertools.SlurSpanner()
    spanner.override.note_head.color = 'red'
    spanner.override.stem.color = 'red'
    spanner.override.beam.color = 'red'
    spanner.override.accidental.color = 'red'
    leaves.set_spanner(spanner)

The example sets a slur on the leaves in divisions 2 - 4 of voice 1
and also apply the overrides listed.

Test and PDF included.

The limitation is that you have to instantiate an actual spanner to get this work.

The mainline previously had an OverrideSpanner that was designed to do exactly
this with no further typographic output.

We should either add the OverrideSpanner back into the mainline or else make Spanner concrete.

It seems better to make Spanner concrete because all spanners have a grob override proxy
(which means that the appearance of an OverrideSpanner class would suggest that some
spanners do not have the ability to override grobs, which is false).