Implemented delayed-evaluation functional mapping, perhaps miraculously.
THIS SLURS ALL LEAVES IN THE SCORE TOGETHER UNDER A SINGLE SLUR (previously existing behavior):

    divisions = score_specification.select_divisions('Voice 1')
    divisions.timespan.select_leaves('Voice 1').set_spanner(spannertools.SlurSpanner())

THIS SLURS EACH DIVISION'S LEAVES TOGETHER USING AS MANY SLURS AS NECESSARY (shiny new behavior):

    divisions = score_specification.select_divisions('Voice 1').map_to_each() # <== special new method
    divisions.timespan.select_leaves('Voice 1').set_spanner(spannertools.SlurSpanner())

Notionally the new functionality says works as type of 'splitter' or 'fork' routine
that forks a single expression into an iterable of arbitrarily many expressions during evaluation.

This means that the second line of the shiny new example should be read as
something like "for each division, go find the leaves that start during the timespan
of the division and then slur them."

Tests and PDFs included.