Activated forbidden_written_duration=None in leaftools leaf-making functions.
Example:

    >>> leaves = leaftools.make_tied_leaf(
    ...     notetools.Note, Duration(9, 8), pitches='C#5',
    ...     forbidden_written_duration=Duration(1, 2))
    >>> staff = Staff(leaves)
    >>> time_signature = contexttools.TimeSignatureMark((9, 8))(staff)

    >>> f(staff)
    \new Staff {
        \time 9/8
        cs''4 ~
        cs''4 ~
        cs''4 ~
        cs''4 ~
        cs''8
    }

The example creates a series of notes totaling 9/8 with the special restriction
that no half notes (or notes of greater written duration) may appear.
This results in the run of quarter notes shown in the example: the four
quarter notes replace the whole note that would normally result from calling the function.

CHANGE: should be included in the 2.12 release notes.