Important optimization.
Made two-line change to Component.stop_duration. 

The relevant change is this:

Previously in Component.py:

    @property
    def stop_offset(self):
        return self.start_offset + self.duration

Now in Component.py:

    @property
    def stop_offset(self):
        self._update_prolated_offset_values_of_entire_score_tree_if_necessary()
        return self._stop

MOTIVATION: things were running very slow while building Gebiete using the
new functionality available in specificationtools: a 44-measure segment
that had been taking 3 seconds to complete rhythm interpreation suddenly 
jumped to 40 seconds to interpet when 10 (or so) leaf set expressions 
(like leaf color, hairpins and pitch) were added.

PROFILE OUTPUT: running iotools.profile_expr() showed that what was taking essentially all of the
time was interpreting the postrhythm parameters (like leaf color, hairpins and pitch, listed
above), which confirmed the observations above. Here's the profiler output for the
40-second interpretation:

Fri Mar  8 19:24:56 2013    _tmp_abj_profile

         52174600 function calls (51659497 primitive calls) in 40.167 seconds

   Ordered by: cumulative time
   List reduced from 729 to 48 due to restriction <48>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000   40.370   40.370 ScoreSpecificationInterface.py:152(interpret)
        1    0.000    0.000   40.370   40.370 ScoreSpecification.py:635(interpret)
        1    0.000    0.000   40.370   40.370 ConcreteInterpreter.py:13(__call__)
        1    0.000    0.000   37.231   37.231 ConcreteInterpreter.py:115(interpret_additional_parameters)
       71    0.107    0.002   36.827    0.519 CounttimeComponentSelectExpression.py:150(evaluate_against_score)
       69    0.000    0.000   23.340    0.338 LeafSetExpression.py:21(_iterate_selected_leaves_in_score)
    28098    0.096    0.000   17.383    0.001 Component.py:185(timespan)
    31625    0.427    0.000   14.075    0.000 TimespanTimespanTimeRelation.py:111(__call__)
   604020    4.226    0.000   12.642    0.000 Duration.py:85(__new__)
  6812135    3.212    0.000   12.308    0.000 {isinstance}
       36    0.004    0.000   12.147    0.337 LeafColorSetExpression.py:11(execute_against_score)
   156007    0.511    0.000   11.921    0.000 Timespan.py:26(__init__)
    28098    0.070    0.000   11.812    0.000 Component.py:173(stop_offset)


STRATEGY: the profiler output makes it look like timespan *creation* takes a bunch of time.
Moreover, reading the CounttimeComponentSelectExpression.evaluate_against_score() implementation
makes it look like timespan *comparison* takes a bunch of time. And what connects timespan creation
to timespan comparison is that Component.timespan is implemented as a property that creates
a new Timespan every time Component.timespan is accessed; that is, timespan comparison
entails timespan creation in the current (one-line) implementation of Component.timespan.
So the strategy was to change Component.timespan from a build-at-every-access property
to a cached-value property (like Component.start_offset).

DISCOVERY: But while reading Component.timespan (and Component.start_offset and Component.stop_offset)
it became apparent that Component.stop_offset could be changed as shown at the start of this commit message.
It didn't seem like that change would make much of a difference. But making the change
took 40-second interpretation of Gebiete down to 29 seconds. Here're the first few lines 
of the new profiler output:

Fri Mar  8 19:49:43 2013    _tmp_abj_profile

         37876876 function calls (37625113 primitive calls) in 29.002 seconds

   Ordered by: cumulative time
   List reduced from 729 to 48 due to restriction <48>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000   29.202   29.202 ScoreSpecificationInterface.py:152(interpret)
        1    0.000    0.000   29.201   29.201 ScoreSpecification.py:635(interpret)
        1    0.000    0.000   29.201   29.201 ConcreteInterpreter.py:13(__call__)
        1    0.000    0.000   26.060   26.060 ConcreteInterpreter.py:115(interpret_additional_parameters)
       71    0.101    0.001   25.654    0.361 CounttimeComponentSelectExpression.py:150(evaluate_against_score)
       69    0.000    0.000   15.844    0.230 LeafSetExpression.py:21(_iterate_selected_leaves_in_score)
    31625    0.416    0.000   13.908    0.000 TimespanTimespanTimeRelation.py:111(__call__)
   156007    0.490    0.000   11.863    0.000 Timespan.py:26(__init__)
   312014    0.764    0.000   10.206    0.000 Timespan.py:621(_initialize_offset)
  4728851    2.200    0.000    9.032    0.000 {isinstance}
   408846    2.798    0.000    8.711    0.000 Duration.py:85(__new__)
    63266    0.095    0.000    8.351    0.000 Timespan.py:610(_get_timespan)
       36    0.004    0.000    8.264    0.230 LeafColorSetExpression.py:11(execute_against_score)

Notice that the calls to Component.stop_offset no longer appear in the profiler output.


CONCLUSION: so the conclusion is that changing Component.stop_offset from an initialize-Duration-and-
add-every-time property to a cached-value property helps enormously when using the new specificationtools
funcationality, probably because counttime component evaluation relies heavily on timespan comparison.

Note, too, that the lean battery of optimization tests housed in the abjad/etc/optimization
directory fails to capture this performance increase. That's fine. But that means that
we'll need to build at least something of a performance battery for specifcationtools
before public release, too.

Also note that further optimization commits are to be expected because specificationtools
interpretation is still taking only 3 seconds for rhythm interpretation but 26 seconds
for postrhythm interpretation. Which is unreasonable.