Optimized Component.timespan management.
The Component.timespan property now simply returns Component._timespan.

This means that Component now comprises a new '_timespan' slot.

This also means that Component._timespan is updated along with
Component._start_offset and Component._stop_offset every time
it is necessary to call offsettools.update_offset_values_of_component().

RESULTS: The first segment of Gebiete previously interpreted in 37.9 million function
calls (taking 30 seconds). That same segment now interprets in 31.1 million function
calls and takes 25 seconds.

Note that is makes sense for Timepsan to implement __slots__.
The reason that this is the case is that there is exactly one timespan
for every component in the system. And if Component implements __slots__
for the reason that there will be potentially many components in a given
score then it makes sense for Timespan to implement __slots__ for the same reason.