Added timerelationtools.get_counttime_components_that_satisfy_time_relation() function.
Example:

    >>> voice = Voice([Note(i % 36, Duration(1, 4)) for i in range(200)])
    >>> timespan_1 = timespantools.Timespan(20, 22)
    >>> time_relation = timerelationtools.timespan_2_starts_during_timespan_1(timespan_1=timespan_1)

    >>> result = timerelationtools.get_counttime_components_that_satisfy_time_relation(
    ...     voice[:], time_relation)

    >>> for counttime_component in result:
    ...     counttime_component
    Note("af'4")
    Note("a'4")
    Note("bf'4")
    Note("b'4")
    Note("c''4")
    Note("cs''4")
    Note("d''4")
    Note("ef''4")

    >>> result.timespan
    Timespan(start_offset=Offset(20, 1), stop_offset=Offset(22, 1))

MOTIVATION: this sort of thing currently runs at O(n) with n equal to the number of input components.
This becomes expensive when this sort of logic is called many times in a row (as happens when 
using specificationtools). So the introduction of this function provides an optimization target.

Note that the example shown here takes 181337 function calls under r9683.