Optimized Duration initialiation.
Duration.__new__() now begins with these three tries:

    try:
        return Fraction.__new__(klass, *args)
    except TypeError:
        pass

    try:
        return Fraction.__new__(klass, *args[0])
    except (AttributeError, TypeError):
        pass

    try:
        return Fraction.__new__(klass, args[0].numerator, args[0].denominator)
    except AttributeError:
        pass

What this does is first try to initialize Duration exactly like Fraction.
Only if that fails does initialization proceed to costlier tests for other type of input pattern.

CHANGE: It is no longer possible to initialize Duratin objects from LilyPond duration strings;
this means that Duration('8..') is no longer possible; use the static method
Duration.from_lilypond_duration_string('8..') instead. This change needs to be noted
in the 2.12 change log.

RESULT: The result of the optimization is significant when thousands of duration are initialized.
A nontrivial specificationtools example decreases from 19.9 million function calls
to 15.9 million function calls with a corresponding decrease in interpretation time. 

RESULT: The optimization also decreases from 105,763 to 78,355 the number of function
calls required to execute the timerelationtools.get_counttime_components_that_satisfy_time_relation()
example in the API that selects from a voice of 200 notes.

NOTE: This change meant updating the parser and reduced ly parser in a couple of places.

NOTE: This change also breaks a MetricalHierarchy.pretty_rtm_format example, too.
That example has been skipped in the doctests temporarily but should be remedied.