Finished integrating bisect into timerelationtools.simple_inequality_to_offset_indices() function.
This ends the first wave of specificationtools optimization.

RESULTS: This wave of optimization included a number of changes to duration
and timespan initialization followed by refactoring the evaluation of
CounttimeComponentSelectExpression objects. This reduced from
84 million (40 seconds) to 11 million (10 seconds) the number of function calls 
necessary to process the part of Gebiete currently taken as canonical. 
This means only 1/8th the functions calls are now required to interpret
the same nontrivial score segment and that interpretation is now 4 times as fast.
The improvements are substantial.

This first wave of specificationtools optimization uncovered a number of to-dos:

TODO #1 (INTERPRETER OPTIMIZATION): The specificationtools interpretation time 
should still be able to be reduced further. While 10 seconds is an acceptable 
amount of time to wait for the interpretation of the 3-page 11x17 colored
score segment used for testing, it seems clear that at least one other structural
change could help significantly. We can identify the change in question
as "select expression caching" and it comes from the observation that realworld
use of specificationtools to build up a segment of nontrivial score will involve
more calls to select_leaves() that to anything else (because all the tweaking of
individual notes, rests, articulations, spanners and so on all involve this operation).
Each call to select_leaves() results in a CounttimeComponentSelectExpression
against which the interpeter eventually calls evaluate_against_score().
The evaluated output of evaluate_against_score() could be stored in memory 
and retrieved using the originating CounttimeComponentSelectExpression as a hash key.
This should provide a performance improvement equal to making each leaf selection
in the score only once; this will not improvement cases of composition in which many
different *unique* leaf selections are made but should improve considerably
compositional cases in which many different *identical* leaf selections are made;
and experience is showing the composition leads to the selection of many of the same
leaf selections over and over (to apply successive parameters or transformations).
So this work can be labeled "(evaluated counttime component) select expression caching".

TODO #2 (FORMAT ENGINE OPTIMIZATION): It seems clear that we can benefit from another
wave of format engine optimization. The various compute times attaching to the recent
optimization target are these:
    Abjad interpretation time:  10 seconds
    Abjad format time:          16 seconds
    LilyPond processing time:   10 seconds
There really is no reason it should take 16 seconds to format the 3-page 11x17 score
taken as the optimization target. First suspicions are the formatting of either
spanners or context marks (or both). It may be time finally to rewrite all or most
of the format engine entirely because, IIRC, the entire approach may be inherently O(n**2)
or O(n**m) in the way that the entire contents of a single spanner may be iterated
many times over right now. The correct implementation of the problem should iterate
the contents of each spanner exactly once.