Streamlined part of contexttools.get_effective_context_marks().
Description: The last part of the function searches all context marks
that govern a component and elects the most recent governing context
mark as the effective context mark. This is perfect for bisect.
So this commit replaces the manual search for most recent context mark
with a call to one of the search methods of SortedCollection.

The speed increase is negligible but the code reads considerably easier
and is now easier to maintain.