It is now possible for spanners to deposit their format contributions at format time.
PURPOSE: This is a fairly complicated commit. The purpose of the commit is to allow
spanners to be migrated over one at a time to a pattern that improves format-time
performance of the system against scores that contain a realworld number of spanners.

SAMPLE RESULT: Consider bechmark spanner score #4 (which features a 200-note voice with slurs
grouping every four notes). The score took ~194,000 function calls to format prior to this commit;
the score now takes ~110,00 function calls to format.

EXPLANATION: Spanners may now optionally implement a _deposit_format_contributions() method.
The purpose of such a method is to allow the spanner to implement an efficient format-time
behavior in which the spanner iterates just those of its leaves to which a format contribution 
is due. In the case of simple slur and beam spanners, for example, it is only
the first and the last of the spanner's leaves that require a format contribution. For scores
that feature something close to a realworld number of spanners this behavior is more
efficient than the old format-time behavior. (The old behavior involved asking every leaf 
in score to dialogue with each of its spanners in search of format contributions.)

As of this commit only SlurSpanner has been migrated to the new behavior.