Increased format speed by 100% with a one-line change. Incredible Python gotcha. Must read.
Particularly unbelievable bug:

When you create a LilyPondFile object with lilypondfiletools.make_basic_lilypond_file()
or equivalent you wind up with a LilyPondFile object that contains a ScoreBlock.
The purpose of ScoreBlock is to act as a (very thin) wrapper around a (potentially very large)
Score object. Now consider the ScoreBlock code that formats the contained Score at format time.

OLD Score._format_pieces (excerpt):

    for x in self:
        if hasattr(x, '_get_format_pieces'):
            result.extend(['\t' + piece for piece in x._get_format_pieces()])
        elif hasattr(x, '_format_pieces'):
            result.extend(['\t' + piece for piece in x._format_pieces])
        elif isinstance(x, str):
            result.append('\t%s' % x)

See the problem that causes a 100% slowdown in format speed??

Scroll down to bypass spoiler alert.























The culprit is -- almost unbelievably -- the ...

    elif hasattr(x, '_format_pieces')

... line.

Why? Because Score._format_pieces (and Note._format_pieces and Rest._format_pieces and so on)
is a *PROPERTY* defined with Python's property decorator. This is of course fine.
But the implementation of Python's built-in hasattr() *EXECUTES 100% OF THE PROPERTY CODE*
in order to make the assessment of whether some object has a property! That's the particularly
unbelievable part.

So this (old) pair of lines ...

    elif hasattr(x, '_format_pieces'):
        result.extend(['\t' + piece for piece in x._format_pieces])

... causes Score._format_pieces (when x was the Score in question, which x usually is)
to be *EXECUTED EXACTLY TWICE*. And this enormously expensive when the score is nontrivial.

What does this mean?

This means that everytime anyone's ever formatted a LilyPondFile object they've waited
exactly twice as long for formatting to complete as was necessary.
        
The fix is incredible easy.

Instead of ...

    elif hasattr(x, '_format_pieces'):
        result.extend(['\t' + piece for piece in x._format_pieces])

... we ask for the equivalent of ...

    try:
        result.extend(['\t' + piece for piece in x._format_pieces])
    except AttributeError:
        pass

... and ensure that x._format_pieces will be executed only once.
        
I elected this as the final form of the entire block:

    for x in self:
        if isinstance(x, str):
            result.append('\t%s' % x)
        elif hasattr(x, '_get_format_pieces'):
            result.extend(['\t' + piece for piece in x._get_format_pieces()])
        else:
            result.extend(['\t' + piece for piece in x._format_pieces])

What's particularly annoying is that I've known about this interaction between
hasattr() and properties for a long time. But this was a particularly old spot
of the code and there was never any reason to suspect anything was going on 
since the functionality was completely correct (even though the performance
was only half of what it should have been).

This almost suggests the need to audit all instances of hasattr() in the system
and go through each by hand to check for the presence of properties.

But just the change committed here should dramatically improve performance
when working with realworld scores (so long as the composer has been embedding
the Score within a LilyPondFile object).

As a point of the comparison the nontrivial segment of Geibete I'm working on
was taking 10.2 million function calls to format. That same segment
now takes only 5.3 million function calls to format, with a corresponding 50%
decrease in total format time.